implement XP gaining
